LPS-98419 Use 'internal' instead of 'impl' for internal classes
